backend/requirements.txt
django-rest-framework
django-cors-headers
psycopg2-binary